pyproject.toml
[project]
name = "piqopiqo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
license = { text = "GPL-3.0-or-later" }
requires-python = ">=3.13"
dependencies = [
    "attrs>=25.4.0",
    "click>=8.3.1",
    "coloredlogs>=15.0.1",
    "pillow>=12.0.0",
    "pyexiftool>=0.5.6",
    "pyobjc-framework-cocoa>=12.1",
    "pyobjc-framework-quartz>=12.1",
    "pyside6>=6.10.1",
]

[project.scripts]
piqopiqo = "piqopiqo.__main__:cli"

# [tool.uv.sources]
# pyqtauto = { path = "/Users/guilhem/Documents/projects/github/pyqtauto" }

[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = ["pytest>=9.0.2"] #"pyqtauto"]
